+++ src/index.js
@@ -12,7 +12,9 @@ import {
   proxyState,
   proxyCompare,
   collectValuables,
+  deproxify,
 } from 'proxyequal';
+
 import { batchedUpdates } from './batchedUpdates';
 
 // global context
@@ -27,12 +29,51 @@ const warningObject = {
 };
 const ReduxStoreContext = createContext(warningObject);
 
+// utils
+
+const shallowEqualDeproxify = (a, b) => {
+  if (a === b) return true;
+  if (typeof a !== 'object' || typeof b !== 'object') return false;
+  const aKeys = Object.keys(a);
+  const bKeys = Object.keys(b);
+  if (aKeys.length !== bKeys.length) return false;
+  return aKeys.every(key => deproxify(a[key]) === deproxify(b[key]));
+};
+
 // helper hooks
 
 const forcedReducer = state => !state;
 const useForceUpdate = () => useReducer(forcedReducer, false)[1];
 
+const useProxyfied = (state) => {
+  // cache
+  const proxyMap = useRef(new WeakMap());
+  const trappedMap = useRef(new WeakMap());
+  // trapped
+  let trapped;
+  if (trappedMap.current.has(state)) {
+    trapped = trappedMap.current.get(state);
+    trapped.reset();
+  } else {
+    trapped = proxyState(state, null, proxyMap.current);
+    trappedMap.current.set(state, trapped);
+  }
+  // update ref
+  const lastProxyfied = useRef(null);
+  useEffect(() => {
+    lastProxyfied.current = {
+      state,
+      affected: collectValuables(trapped.affected),
+    };
+  });
+  return {
+    proxyfiedState: trapped.state,
+    lastProxyfied,
+  };
+};
+
 // patch store with batchedUpdates
+
 const patchReduxStore = (origStore) => {
   if (!batchedUpdates) return origStore;
   const listeners = [];
@@ -98,34 +139,19 @@ export const useReduxDispatch = () => {
 
 export const useReduxState = () => {
   const forceUpdate = useForceUpdate();
-  // store&state
+  // redux store
   const store = useContext(ReduxStoreContext);
+  // redux state
   const state = store.getState();
-  // trapped
-  const proxyMap = useRef(new WeakMap());
-  const trappedMap = useRef(new WeakMap());
-  let trapped;
-  if (trappedMap.current.has(state)) {
-    trapped = trappedMap.current.get(state);
-    trapped.reset();
-  } else {
-    trapped = proxyState(state, null, proxyMap.current);
-    trappedMap.current.set(state, trapped);
-  }
-  // update refs
-  const lastState = useRef(null);
-  const lastAffected = useRef(null);
-  useEffect(() => {
-    lastState.current = state;
-    lastAffected.current = collectValuables(trapped.affected);
-  });
+  // proxyfied
+  const { proxyfiedState, lastProxyfied } = useProxyfied(state);
   // subscription
   useEffect(() => {
     const callback = () => {
       const changed = !proxyCompare(
-        lastState.current,
+        lastProxyfied.current.state,
         store.getState(),
-        lastAffected.current,
+        lastProxyfied.current.affected,
       );
       drainDifference();
       if (changed) {
@@ -136,8 +162,56 @@ export const useReduxState = () => {
     callback();
     const unsubscribe = store.subscribe(callback);
     return unsubscribe;
-  }, [store, forceUpdate]);
-  return trapped.state;
+  }, [store]); // eslint-disable-line react-hooks/exhaustive-deps
+  return proxyfiedState;
+};
+
+export const useReduxStateMapped = (mapState) => {
+  const forceUpdate = useForceUpdate();
+  // redux store
+  const store = useContext(ReduxStoreContext);
+  // redux state
+  const state = store.getState();
+  // proxyfied
+  const { proxyfiedState, lastProxyfied } = useProxyfied(state);
+  // mapped
+  const mapped = mapState(proxyfiedState);
+  // update ref
+  const lastMapped = useRef(null);
+  useEffect(() => {
+    lastMapped.current = {
+      mapped,
+      mapState,
+    };
+  });
+  // subscription
+  useEffect(() => {
+    const callback = () => {
+      let changed = !proxyCompare(
+        lastProxyfied.current.state,
+        store.getState(),
+        lastProxyfied.current.affected,
+      );
+      drainDifference();
+      if (!changed) return; // no state parts interested are changed.
+      try {
+        changed = !shallowEqualDeproxify(
+          lastMapped.current.mapped,
+          lastMapped.current.mapState(store.getState()),
+        );
+      } catch (e) {
+        changed = true; // props are likely to be updated
+      }
+      if (changed) {
+        forceUpdate();
+      }
+    };
+    // run once in case the state is already changed
+    callback();
+    const unsubscribe = store.subscribe(callback);
+    return unsubscribe;
+  }, [store]); // eslint-disable-line react-hooks/exhaustive-deps
+  return mapped;
 };
 
 export const useReduxStateSimple = () => {
@@ -173,6 +247,6 @@ export const useReduxStateSimple = () => {
       used.current = {};
     };
     return cleanup;
-  }, [store, forceUpdate]);
+  }, [store]); // eslint-disable-line react-hooks/exhaustive-deps
   return new Proxy(state, handler);
 };